requirements.txt
# Core dependencies for Granite 4.0-1B finetuning
# Install: pip install -r requirements.txt

# LLaMA-Factory
llamafactory>=0.9.0

# Transformers ecosystem
transformers>=4.45.0
datasets>=2.14.0
accelerate>=0.27.0
peft>=0.10.0
trl>=0.8.0
bitsandbytes>=0.43.0

# Tokenizers
sentencepiece
tiktoken
protobuf

# API client for data generation
openai>=1.0.0
httpx

# Utilities
tqdm
pandas
numpy
python-dotenv
pyyaml
fire

# Evaluation metrics
rouge-score
sacrebleu
bert-score
jieba
scikit-learn
